+++ lua/monokai-v2/theme/plugins/copilot.lua
@@ -1,0 +1,11 @@
+local M = {}
+
+--- @param c Colorscheme The color palette
+function M.get(c, _, _)
+  return {
+    CopilotSuggestion = { fg = c.base.dimmed1, italic = true },
+    CopilotAnnotation = { fg = c.base.dimmed1, italic = true },
+  }
+end
+
+return M