+++ lua/monokai-v2/theme/plugins/lsp_signature.lua
@@ -1,0 +1,22 @@
+local M = {}
+
+--- @param c Colorscheme The color palette
+--- @param config Config
+--- @param hp Helper
+function M.get(c, config, hp)
+  local isBackgroundClear = vim.tbl_contains(config.background_clear, "float_win")
+  local transparent_bg = c.base.black
+
+  return {
+    LspSignatureActiveParameter = {
+      fg = c.base.orange,
+      bold = true,
+      underline = true,
+    },
+    LspSignatureHint = {
+      fg = c.base.dimmed1,
+    },
+  }
+end
+
+return M